requirements.txt
deezloader
python-telegram-bot
git+git://github.com/pyrogram/pyrogram@master
tgcrypto
python-acrcloud
libtmux
pymongo[srv]==3.12.0